authorization method changed

+++ src/service/EmployeeService.js
@@ -2,34 +2,34 @@ import apiClient from '../client/apiClient';
 
 const EmployeeService = {
 
-  async getEmployees() {
-    const response = await apiClient.get('/api/v1/emp/employees', {withCredentials: true});
+  async getEmployees(token) {
+    const response = await apiClient.get('/api/v1/emp/employees', {headers: {Authorization: `Bearer ${token}`}} ,{withCredentials: true});
     return response;
   },
 
-  async getEmployee(id){
-    const response = await apiClient.get(`/api/v1/emp/employees/${id}`)
+  async getEmployee(id, token){
+    const response = await apiClient.get(`/api/v1/emp/employees/${id}`, {headers: {Authorization: `Bearer ${token}`}} ,{withCredentials: true})
     return response;
   },
 
 
-  async createEmployees(payload){
-    const response = await apiClient.post('/api/v1/emp/employees', payload);
+  async createEmployees(payload, token){
+    const response = await apiClient.post('/api/v1/emp/employees', payload, {headers: {Authorization: `Bearer ${token}`}} ,{withCredentials: true});
     return response;
   },
 
-  async updateEmployees(id, payload){
-    const response = await apiClient.put(`/api/v1/emp/employees/${id}`, payload)
+  async updateEmployees(id, payload, token){
+    const response = await apiClient.put(`/api/v1/emp/employees/${id}`, payload, {headers: {Authorization: `Bearer ${token}`}} ,{withCredentials: true})
     return response;
   },
 
-  async searchEmployees(query) {
-    const response = await apiClient.get(`/api/v1/emp/search/${query}`);
+  async searchEmployees(query, token) {
+    const response = await apiClient.get(`/api/v1/emp/search/${query}`, {headers: {Authorization: `Bearer ${token}`}} ,{withCredentials: true});
     return response;
   },
 
-  async deleteEmployee(eid) {
-    const response = await apiClient.delete(`/api/v1/emp/employees?eid=${eid}`);
+  async deleteEmployee(eid, token) {
+    const response = await apiClient.delete(`/api/v1/emp/employees?eid=${eid}`, {headers: {Authorization: `Bearer ${token}`}} ,{withCredentials: true});
     return response.data;
   },
